server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Server for the project"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "colorlog>=6.9.0",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "lightgbm>=4.6.0",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
